+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Any `DateInput` values entered that exceed any specified max/minDate will now be reset to null,
   instead of being set to the boundary date (which was surprising and potentially much less obvious
   to a user that their input had been adjusted automatically).
+* New `Clock` component has been added for displaying the current time for a given timezone.
 
 ### 🐞 Bug Fixes